requirements.txt
# Core ML & Data
numpy<2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow-cpu>=2.14.0
joblib>=1.3.0
openpyxl>=3.1.2     # REQUIRED for Indo Hate Excel dataset

# NLP & Text
nltk>=3.8.1
regex>=2023.0.0
transformers>=4.30.0
tf-keras

# UI & Visualization
nicegui>=1.4.0
plotly>=5.17.0
matplotlib>=3.8.0
seaborn

# Utilities
tqdm

# Backend
flask
flask-cors
